Update README.md
This Sentiment Analysis project aims to determine the sentiment of a given piece of text—whether it expresses a positive, negative, or neutral emotion. Sentiment analysis is a core task in Natural Language Processing (NLP) and plays a key role in understanding customer feedback, social media opinions, product reviews, and more.

In this project, a machine learning-based approach is used to classify the sentiment of the text, enabling automated understanding of the emotional tone behind the words. The sentiment classification can be applied to various domains, including marketing, customer service, and social media monitoring.